zrocks: Delete/Trim mapping pieces / Garbage collection
User buffers are mapped into zone in multiple pieces. When a buffer
is deleted by the appilication, it must call zrocks_trim for each
piece of the buffer. This will trigger a process that frees the
zones linked to the deleted buffer. This is the zone garbage
collection.

If the application (e.g. RocksDB) called zrocks_write with a correct
level (e.g. RocksDB level), then the write-amplification should
be close to 1 when the buffers are deleted (no data must be moved
before the zone is freed).

Signed-off-by: Ivan L. Picoli <i.picoli@samsung.com>